rename intl context file
